requirements.txt
blurhash==1.1.4
certifi==2025.1.31
charset-normalizer==3.4.1
decorator==5.2.1
feedparser==6.0.11
idna==3.10
Mastodon.py==2.0.1
python-dateutil==2.9.0.post0
python-magic==0.4.27
PyYAML==6.0.2
requests==2.32.3
sgmllib3k==1.0.0
six==1.17.0
urllib3==2.3.0
